AdvancedCalculatorWithMethodOverloading
In this project, we will create an advanced calculator using method overloading in Java. Our calculator will perform addition, subtraction, multiplication and division operations with different parameter types and number quantities. In this way, the user can enter two or three numbers for the operations and the methods with the same name will perform the correct operation according to the number of parameters entered.